SERVER and CLIENT/ add exp seed data, fix home page to display exp data, add routes to fetch by order, tag, user
