+++ server/app.ts
@@ -26,7 +26,7 @@ import ImageRouter from './routes/PhotoUpload';
 import ParadesRoutes from './routes/Parades';
 import GigsRoutes from './routes/ScrapeEvents';
 import MailListRoutes from './routes/MailList';
-import TestRoutes from './routes/Test';
+import ContentRoutes from './routes/Content';
 import { User } from './db/index';
 
 // import { Sequelize } from "sequelize";
@@ -60,7 +60,7 @@ app.use('/api/images', ImageRouter);
 app.use('/api/parades', ParadesRoutes);
 app.use('/api/gigs', GigsRoutes);
 app.use('/api/mail', MailListRoutes);
-app.use('/api/test', TestRoutes);
+app.use('/api/content', ContentRoutes);
 app.use(
   cors({
     origin: ['http://localhost:4000'],
